# dev/backend/docs/openapi/condohub.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CondoHub Backend API",
    "version": "1.0.0",
    "description": "API do CondoHub Backend (Dev 1 + Dev 2, refletindo endpoints implementados)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://api-dev.condohub.local",
      "description": "Exemplo ambiente dev"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth Global"
        ],
        "summary": "Global tenant login",
        "description": "Rota pública global. Quando email e senha identificam uma única conta tenant, retorna AuthResponse. Quando identificam múltiplas contas tenant, retorna InstanceSelectionRequiredResponse para seleção de condomínio.",
        "operationId": "globalLogin",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated directly or instance selection required",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuthResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InstanceSelectionRequiredResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/InvalidCredentialsError"
          },
          "403": {
            "$ref": "#/components/responses/InstanceSuspendedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
...
      "ValidationError": {
        "description": "Validation error (code: VALIDATION_ERROR)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "VALIDATION_ERROR",
              "message": "Validation failed",
              "details": {}
            }
          }
        }
      },
...
      "InstanceSuspendedError": {
        "description": "Instance suspended (code: INST_SUSPENDED)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "INST_SUSPENDED",
              "message": "This instance is suspended",
              "details": {}
            }
          }
        }
      },
      "InvalidCredentialsError": {
        "description": "Invalid login credentials (code: INVALID_CREDENTIALS)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "INVALID_CREDENTIALS",
              "message": "Invalid credentials",
              "details": {}
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "code",
          "message",
          "details"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "example": {
          "error": true,
          "code": "SOME_CODE",
          "message": "Human message",
          "details": {}
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN_GLOBAL",
          "SINDICO_ADMIN",
          "FUNC_ENTREGAS",
          "FUNC_MANUTENCAO",
          "MORADOR"
        ]
      },
      "AuthUser": {
        "type": "object",
        "required": [
          "id",
          "instanceId",
          "unitId",
          "roles",
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "instanceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "instanceKey": {
            "type": "string",
            "nullable": true,
            "description": "Chave pública da instância tenant. Retornada em logins tenant/global quando aplicável; ausente ou null para admin global."
          },
          "unitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "expiresInSec",
          "user"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresInSec": {
            "type": "integer",
            "enum": [
              300
            ]
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
      "GlobalLoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "InstanceSelectionOption": {
        "type": "object",
        "required": [
          "instanceId",
          "instanceKey",
          "instanceName",
          "userId",
          "unitId",
          "unitLabel",
          "roles"
        ],
        "properties": {
          "instanceId": {
            "type": "string",
            "format": "uuid"
          },
          "instanceKey": {
            "type": "string"
          },
          "instanceName": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "unitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "unitLabel": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "InstanceSelectionRequiredResponse": {
        "type": "object",
        "required": [
          "requiresInstanceSelection",
          "selectionToken",
          "options"
        ],
        "properties": {
          "requiresInstanceSelection": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "selectionToken": {
            "type": "string",
            "description": "JWT temporário assinado com purpose instance_selection, expira em 5 minutos e contém allowed: [{ userId, instanceId }]."
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstanceSelectionOption"
            }
          }
        }
      },
...
    }
  }
}